Конвертация ЭПД Модуля согласования Изменил конвертацию с int на double?         public double? weight_place_br { get; set; }	//	Стандартна маса одного місця брутто (кг)         public double? weight_place_net { get; set; }	//	Стандартна маса одного місця нетто (кг)
1.20.0403.0
- Внесено зміни в схеми ЕПД версії 1.11 та 1.10.
	- в тегі otpr значення weight_place_br змінено на формат N(5.1)
       	- в тегі otpr значення weight_place_net змінено на формат N(5.1)

+++ UZ_SMS/UZ_Convert.cs
@@ -335,8 +335,8 @@ namespace UZ
     public class PAC
     {
         public int? kol_pac { get; set; }	//	Кількість місць упаковки
-        public int? weight_place_br { get; set; }	//	Стандартна маса одного місця брутто (кг)
-        public int? weight_place_net { get; set; }	//	Стандартна маса одного місця нетто (кг)
+        public double? weight_place_br { get; set; }	//	Стандартна маса одного місця брутто (кг)
+        public double? weight_place_net { get; set; }	//	Стандартна маса одного місця нетто (кг)
     }
 
     public class PASS_MARK
@@ -1234,8 +1234,8 @@ namespace UZ
             if (node.Attributes.Count > 0)
             {
                 tag.kol_pac = getAttributes<int?>(node, "kol_pac");
-                tag.weight_place_br = getAttributes<int?>(node, "weight_place_br");
-                tag.weight_place_net = getAttributes<int?>(node, "weight_place_net");
+                tag.weight_place_br = getAttributes<double?>(node, "weight_place_br");
+                tag.weight_place_net = getAttributes<double?>(node, "weight_place_net");
             }
         }
         /// <summary>
